use latest isar and it work on 32 bit devices too
Former-commit-id: c5b98266bd7becb6ecf560322658da0083c2095f

+++ packages/flipper_models/lib/isar/discount.dart
@@ -11,7 +11,7 @@ class Discount {
   late String name;
 
   /// this is to get ready when there will be option for percentage instead of amount!
-  late int? amount;
+  late double? amount;
   @Index()
   late int branchId;
